[EPMDEDP-7105]: refactor: Ensure we are aligned with SSO enabled flag
We are enabling SSO for admin console through the flag, but
at the same time - disable SSO for admin console deployment.
admin-cosnole-operator will update deployment template later
properly (so technically there are no issues).

But this change addresses Out-of-Sync issue for
GitOps approach when we expect code is equal to state

Signed-off-by: Sergiy Kulanov <sergiy_kulanov@epam.com>
Change-Id: If0545c932087711753a1ddc0aacb41bafe885b4c